pyproject.toml
[project]
name = "vrptw-route-planner"
version = "0.1.0"
description = "A PyQt6-based application for solving Vehicle Routing Problems with Time Windows (VRPTW)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "PyQt6>=6.6.1",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-qt>=4.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ui", "models"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
